requirements.txt
aniso8601==9.0.1
bcrypt==4.2.0
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.0.2
pandas==2.2.3
PyJWT==2.9.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
